rip: Fix build on windows by reverting a few fragments of 67b7092.
emit()-ing a size_t without casting works on linux and macos,
because it's the same size there as an unsigned long - 64 bits.
On windows, size_t is 64 bits, unsigned long is 32 bits, and
emit() doesn't have an overload for the 64 bit unsigned long long.

+++ src/inet/routing/rip/Rip.cc
@@ -287,7 +287,7 @@ RipRoute *Rip::importRoute(IRoute *route, RipRoute::RouteType type, int metric, 
     }
 
     ripRoutes.push_back(ripRoute);
-    emit(numRoutesSignal, ripRoutes.size());
+    emit(numRoutesSignal, (unsigned long)ripRoutes.size());
     return ripRoute;
 }
 
@@ -905,7 +905,7 @@ void Rip::addRoute(const L3Address& dest, int prefixLength, const InterfaceEntry
     ripRoute->setLastUpdateTime(simTime());
     ripRoute->setChanged(true);
     ripRoutes.push_back(ripRoute);
-    emit(numRoutesSignal, ripRoutes.size());
+    emit(numRoutesSignal, (unsigned long)ripRoutes.size());
     triggerUpdate();
 }
 
@@ -1045,7 +1045,7 @@ void Rip::purgeRoute(RipRoute *ripRoute)
     remove(ripRoutes, ripRoute);
     delete ripRoute;
 
-    emit(numRoutesSignal, ripRoutes.size());
+    emit(numRoutesSignal, (unsigned long)ripRoutes.size());
 }
 
 /**
@@ -1141,7 +1141,7 @@ void Rip::deleteInterface(const InterfaceEntry *ie)
             it++;
     }
     if (emitNumRoutesSignal)
-        emit(numRoutesSignal, ripRoutes.size());
+        emit(numRoutesSignal, (unsigned long)ripRoutes.size());
 }
 
 int Rip::getInterfaceMetric(InterfaceEntry *ie)
